Remove fake element created before links so background aligns correctly
This rule seems to no longer apply to the skin in it's current state,
and seems to be causing problems with references. These rules are an issue
for that it creates a fake block element (which adds extra spacing at the
top), creates a static height, and pulls it up with a negative top margin.

It seems best to remove this CSS block.

Bug: T181490
Change-Id: Id109e42731202f0c6e1ef6ff8855617e0795b35d