Make Timeless aware of OOUI multiselect widgets
These OOUI widgets are special in so far that:

* OOUI makes the actual <input> element invisible and partly fails
  to do so because Timeless sets properties on <input>s that are not
  reset by OOUI.

* An outer <div> is styled like an <input> instead. Timeless is not
  aware of this and forgets to set it's overrides for <input>s there.

This patch is motivated by the AdvancedSearch extension which uses
these widgets, but is not limited to this.

This patch also does not fix all compatibility issues between Timeless
and these OOUI widgets, but greatly reduces the main issues. More to
come in later patches.

Bug: T192303
Change-Id: I8a4fc3f74dabdb54fd95cd133bb4fe28944c4a40