Display notifications in echo callout for mobile view
Override the inline style `width: 0px` on .oo-ui-clippableElement-clippable
elements within .mw-echo-ui-overlay.

Also adjust some other styles for the callout, so that it stays within the
screen width, but is not any wider than it would be otherwise (500px).
Doing so causes the anchor triangle at the top to not align with the
relevant notifications icon, so it is hidden with `display:none`.
(On a narrow mobile screen, it doesn't seem to serve much purpose anyway.)

Bug: T197064
Change-Id: Iaedde85b5c31ab244d4f72c54c9aad6a1b4ebb73