Fix display of categories in the sidebar
Switching from a unique approach -- getting category names, querying
the database, and then building links -- to the standard
OutputPage::getCategoryLinks() fixes the display of redlinks (T175713),
display order (T197852), and perhaps other bugs in the subtasks
of T160645.

A downside to this approach is that categories will not be shown on
history pages, but this is consistent with other skins (e.g. Vector
and Monobook) not displaying categories on history pages.

Bug: T186089
Bug: T175716
Bug: T175714
Bug: T175713
Bug: T197852
Bug: T160645
Change-Id: Ifad999033e8448615128cba4d6386045a9fc749b