Manually optimize all SVG images
I applied a series of manual optimizations, most notably:
* Remove unused IDs.
* Remove attributes with standard values, e.g. opacity="1" is the
  default anyway.
* Remove attributes that don't have an effect, e.g. font styling
  where no fonts are used, or stroke attributes when there is no
  stroke anyway.
* Replace millimeters with pixels.
* I also tried to remove most translations. A lot of files had all
  their coordinates offset by -1032.362. I gave up for a few files.
  It might be worth looking into these later.
* brackets-grey.svg not only had a translation applied, but a
  non-standard viewBox as well. I wrote a tiny script to re-calculate
  all the numbers in this file. As a result it's much more crisp now.

Note that some of these optimizations might alter the appearence of
the image. I mean, they shouldn't, but better check each file manually
before merging this patch.

I did not touched the two files "cat-grey.svg" (I assume this is a
placeholder) as well as "editing icons.svg" (seems to be a temporary
file that should be removed).

Change-Id: Ie2dadef9d5979b563aa0ed0f5d8a69c534b76a4e