Custom logo rendering, including text wordmark images
Use a file upload onwiki for simplest approach, or specify srcset info in an
array if that's not an option or otherwise not desired.

For instance to just reuse a $wgLogo and $wgLogoHD already set for a 160x160px
logo, but where the HD versions actually come out the right size:
$wgTimelessLogo = [
	'1x' => $wgLogo,
	'1.5x' => $wgLogoHD['1.5x']",
	'2x' => $wgLogoHD['2x'],
	'width' => 160,
	'height' => 160
];

Or to use File:Cows.svg uploaded onwiki:
$wgTimelessLogo = 'Cows.svg';

And the same thing with $wgTimelessWordmark to replace the site title string
with an image version.

Fallback behaviour remains unchanged, if just not using these.

Bug: T132004
Bug: T230663
Change-Id: I0d7b398935045ba69462252d40832ed95ad2d5ed