Merge 'skins.timeless.misc' back into 'skins.timeless'
Follows-up f450f7e13 and 1bb12ffaae85a1 which introduce and
explaine that module's need based on what I believe was a
misunderstanding.

Specifying an options object (e.g. with a "media" key) is optional
for a stylesheet, and a module can have any number of CSS or LESS
files that are packaged together. The syntax for this was designed
for PHP and unfortunately does not translate elegantly to JSON,
but it does work. In PHP, examples of this can be found that look
as follows:

 'styles' => [
   'foo.css',
   'bar.css' => [ 'media' => 'print' ],
   'quux.css',
 ],

Other examples of this in JSON can be found in Nostalgia.git
and ArticlePlaceholder.git.

Change-Id: Ie2a39a79c0892a1c5699a05071216eb5fe660bc1