Render unnecessarily large 2x logos to ensure we support unnecessarily high-res devices
Only does this with wordmarks; desktop devices are a lot less likely to merit this, and
the larger desktop logos are probably a rather bad idea to make unnecessarily large
anyway. Whereas the wordmarks should still be pretty small in terms of filesize as
they're going to be pretty simple, and are all we normally show on mobile.

Bug: T232540
Change-Id: I1932a99d8784eadf02f2711d9557425de0db0100