Scrollbars on wikitables and junk
It's called shitty js. Affordance to label overflow is slightly wrong
for current behaviour, as it basically presents as a toggle. Maybe it
should toggle something! Like popouts, that would be fun. I'unno.

Slightly more framing on wikitables etc, but apply to all tables,
because why not.

Rejig to come with actual popouts or stuff...

Bug: T179265
Bug: T265785
Change-Id: I607cb87687fcbcb481dc2fb7165c52b9e9c35873