resources: increase specificity of mw-editfont font-family in forms.less
In certain conditions, notably when using ?action=edit&section=new with
CodeMirror in the 2017 Wikitext Editor, the .mw-editfont font-family
rules were not being applied as intended. This caused some minor visual
inconsistencies. Instead, the rules from core were applied because they
were loaded later. This patch increases the specificity of the selectors
in forms.less so they are applied instead.

Bug: T271177
Change-Id: I9d1b62baa934f738d1c52972cfaf3339c2d1edce