Timeless should only generate body of page
In I203881d7ab035bc953d113106f21348187b51374 a bodyOnly skin
option was added to allow skins to be limited to generation of the
body tag contents. When used, generation of html, body, and head is
deferred to OutputPage.

Since Timeless is deployed on Wikimedia wikis, this is preferred
as it guarantees that debugging information, and head
tags are consistent for all Wikimedia deployed skins.

Bug: T62846
Change-Id: I3353f8b63e8acf3ab29d8c24dd868ad3f121e5bb