Add style exception for OOUI buttons' margin
On OOUI forms the text for the checkboxes tended to be rather
misaligned, with the text usually ending up getting shunted off to the
bottom-right corner. It seems the margin being added by the special-case
in this rule was responsible, so I added an exception to it. In testing
things seem to look okay.

Change-Id: Id3ca935e5c83d7c96edcb02d1c6fd72af8e81e58