Yoink 'content-media' feature from core
Unfortunately, that seems to be what this skin is doing for other files.
The styles are then adapted to match the theme.  If would be much nicer
to just enable the feature but that pulls in content.thumbnails-* which
this skin has duplicated and adapted.

The resources added here are only needed if
$wgParserEnableLegacyMediaDOM is `false` though, but hopefully that'll
soon be the default.

Alas, when VE loads, it pulls in "mediawiki.skinning.content.parsoid"
which overwrites some of the styles adapted here.

Bug: T287965
Change-Id: Ia9278b4c164ea1ddeca49094c572df70e660043e