Remove unnecessary CodeMirror skin styles
The code being removed was added with I8142c5055348. It seems to
unconditionally make the CodeMirror editor monospaced, when it should
respect the 'editfont' user preference. T247325 suggests respecting the
font preference somehow worked, but at any rate it doesn't anymore so we
simply remove the styles.

Going by the 'editfont' preference is fixed in CodeMirror 6 with
Iaaf65e47ce8ed.

The ext.CodeMirror.visualEditor skin style remains, as it's necessary to
have monospaced font in the 2017 editor (T245568#5897582).

Bug: T357794
Bug: T247325
Bug: T245568
Change-Id: I77765804fd777709d0d60f0cba7ec6d9bf898671