Do not disable Parsoid styles
Generally this is a bad idea as it is important for supporting
Parsoid roll out especially for skins in Wikimedia production.

This feature was also renamed content-body so the name being
passed might be removed at any point

Styles are removed from shared.css that are present in the content-body
feature to avoid duplication.

Follows-Up: I2d600a7d9768942b331f0de1dd1183e438144ad5
Change-Id: Ieae29df9be5aa1a38afaf4447bca45827c3148e2