tables: use offsetWidth when adding .overflowed
There’s an odd discrepancy with how jQuery calculates the width of
tables on macOS that causes a small difference in width making us think
a table is overflowing when it’s in fact fitting perfectly, resulting in
a scrollbar for the table being added when it isn’t necessary. Using the
browser’s native .offsetWidth instead of jQuery’s .outerWidth() seems to
resolve this issue.

Bug: T402848
Change-Id: I43bdfe79ded6854999a1fcc15efe59c35a0cb810