talk pages: decrease <dd> indentation on mobile
People don't want to outdent, so best we can do is indent nested <dd>
tags less. It's still possible for this to get silly, but it's much less
likely. (The 'ideal' solution would be some sort of continuous decrease
in indentation the deeper you go that's asymptotic to a margin that text
is still readable at. That'd be hard though.)

Bug: T369397
Change-Id: Icbe7989c9803458aa7940af19eab801ce6cdc348